requirements.txt
azure-functions
azure-cosmos
python-dotenv
